requirements.txt
Click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
certifi==2021.10.8
charset-normalizer==2.0.9
mysql-connector-python==8.0.27
requests==2.26.0
six==1.16.0
PyMySQL==1.0.2
g
redirect
render_template
request 
session
abort
jsonify
url_for
pyautogui


